Changed HashMap Key to Integer

+++ src/planespotter/dataclasses/Flight.java
@@ -14,7 +14,7 @@ public class Flight {
     private String callsign;
     private String flightnr;
     private Plane plane;
-    private HashMap<Long, DataPoint> datapoints;
+    private HashMap<Integer, DataPoint> datapoints;
 
     /**
      * Constructor
@@ -27,7 +27,7 @@ public class Flight {
      * @param flightnr String Flightnumber, not the FlightID
      * @param datapoints HashMap<Long, DataPoint> Hashmap containing all Datapoint keyed by Timestamp
      */
-    public Flight (int id, Airport start, Airport dest, String callsign, Plane plane, String flightnr, HashMap<Long, DataPoint> datapoints) {
+    public Flight (int id, Airport start, Airport dest, String callsign, Plane plane, String flightnr, HashMap<Integer, DataPoint> datapoints) {
         this.id = id;
         this.start = start;
         this.dest = dest;
@@ -86,7 +86,7 @@ public class Flight {
     /**
      * @return HashMap<Long, DataPoint> All of the Tracking Data of the Flight
      */
-    public HashMap<Long, DataPoint> getDataPoints() {
+    public HashMap<Integer, DataPoint> getDataPoints() {
     	return this.datapoints;
     }
 }
